backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
requests
python-jose[cryptography]
python-dotenv
psycopg2-binary
boto3